+++ src/main/kotlin/com/wutsi/platform/core/logging/servlet/KVLoggerFilter.kt
@@ -65,5 +65,8 @@ class KVLoggerFilter(
 
         val params = request.parameterMap
         params.keys.forEach { kv.add("http_param_$it", params[it]?.toList()) }
+
+        request.getHeader("Authorization")?.let { kv.add("authorization", "********") }
+        request.getHeader("X-Api-Key")?.let { kv.add("api_key", "********") }
     }
 }