+++ src/main/kotlin/com/wutsi/platform/core/security/ApiKeyProvider.kt
@@ -1,0 +1,5 @@
+package com.wutsi.platform.core.security
+
+interface ApiKeyProvider {
+    fun getApiKey(): String?
+}