+++ src/main/kotlin/com/wutsi/platform/core/security/spring/SecurityConfigurationNone.kt
@@ -2,6 +2,7 @@ package com.wutsi.platform.core.security.spring
 
 import org.slf4j.LoggerFactory
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty
+import org.springframework.context.ApplicationContext
 import org.springframework.context.annotation.Configuration
 import org.springframework.security.config.annotation.web.builders.HttpSecurity
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity
@@ -12,7 +13,9 @@ import org.springframework.security.config.annotation.web.configuration.EnableWe
     value = ["wutsi.platform.security.type"],
     havingValue = "none"
 )
-open class SecurityConfigurationNone : AbstractWebSecurityConfiguration() {
+open class SecurityConfigurationNone(
+    context: ApplicationContext
+) : AbstractWebSecurityConfiguration(context) {
     companion object {
         private val LOGGER = LoggerFactory.getLogger(SecurityConfigurationNone::class.java)
     }