+++ src/main/kotlin/com/wutsi/platform/core/security/spring/SpringApiKeyRequestInterceptor.kt
@@ -1,19 +1,19 @@
 package com.wutsi.platform.core.security.spring
 
-import com.wutsi.platform.core.security.TokenProvider
+import com.wutsi.platform.core.security.ApiKeyProvider
 import org.springframework.http.HttpRequest
 import org.springframework.http.client.ClientHttpRequestExecution
 import org.springframework.http.client.ClientHttpRequestInterceptor
 import org.springframework.http.client.ClientHttpResponse
 
-class SpringAuthorizationRequestInterceptor(
-    private val tokenProvider: TokenProvider
+class SpringApiKeyRequestInterceptor(
+    private val apiKeyProvider: ApiKeyProvider
 ) : ClientHttpRequestInterceptor {
 
     override fun intercept(request: HttpRequest, body: ByteArray, exec: ClientHttpRequestExecution): ClientHttpResponse {
-        val token = tokenProvider.getToken()
-        if (token != null)
-            request.headers["Authorization"] = listOf("Bearer $token")
+        val apiKey = apiKeyProvider.getApiKey()
+        if (apiKey != null)
+            request.headers["X-Api-Key"] = listOf(apiKey)
 
         return exec.execute(request, body)
     }