+++ src/main/kotlin/com/wutsi/platform/core/test/TestApiKeyProvider.kt
@@ -1,0 +1,7 @@
+package com.wutsi.platform.core.test
+
+import com.wutsi.platform.core.security.ApiKeyProvider
+
+class TestApiKeyProvider(private var apiKey: String? = null) : ApiKeyProvider {
+    override fun getApiKey() = apiKey
+}