+++ src/main/kotlin/com/wutsi/platform/core/test/TestTokenProvider.kt
@@ -2,7 +2,6 @@ package com.wutsi.platform.core.test
 
 import com.wutsi.platform.core.security.TokenProvider
 
-class TestTokenProvider(private var token: String) : TokenProvider {
-    override fun getToken(): String =
-        token
+class TestTokenProvider(private var token: String? = null) : TokenProvider {
+    override fun getToken() = token
 }