+++ src/test/kotlin/com/wutsi/platform/core/security/spring/SpringApiKeyRequestInterceptorTest.kt
@@ -4,23 +4,21 @@ import com.nhaarman.mockitokotlin2.any
 import com.nhaarman.mockitokotlin2.doReturn
 import com.nhaarman.mockitokotlin2.mock
 import com.nhaarman.mockitokotlin2.whenever
-import com.wutsi.platform.core.security.TokenProvider
+import com.wutsi.platform.core.security.ApiKeyProvider
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.springframework.http.HttpHeaders
 import org.springframework.http.HttpRequest
 import org.springframework.http.client.ClientHttpRequestExecution
 import org.springframework.http.client.ClientHttpResponse
-import kotlin.test.assertNull
-import kotlin.test.assertTrue
 
-internal class SpringAuthorizationRequestInterceptorTest {
+internal class SpringApiKeyRequestInterceptorTest {
     private lateinit var headers: HttpHeaders
     private lateinit var request: HttpRequest
     private lateinit var response: ClientHttpResponse
-    private lateinit var tokenProvider: TokenProvider
     private lateinit var exec: ClientHttpRequestExecution
-    private lateinit var interceptor: SpringAuthorizationRequestInterceptor
+    private lateinit var apiKeyProvider: ApiKeyProvider
+    private lateinit var interceptor: SpringApiKeyRequestInterceptor
 
     @BeforeEach
     fun setUp() {
@@ -32,26 +30,26 @@ internal class SpringAuthorizationRequestInterceptorTest {
         exec = mock()
         doReturn(response).whenever(exec).execute(any(), any())
 
-        tokenProvider = mock()
+        apiKeyProvider = mock()
 
-        interceptor = SpringAuthorizationRequestInterceptor(tokenProvider)
+        interceptor = SpringApiKeyRequestInterceptor(apiKeyProvider)
     }
 
     @Test
-    fun interceptWithToken() {
-        doReturn("foo").whenever(tokenProvider).getToken()
+    fun interceptWithApiKey() {
+        doReturn("foo").whenever(apiKeyProvider).getApiKey()
 
         interceptor.intercept(request, ByteArray(10), exec)
 
-        assertTrue(headers["Authorization"]!![0]!!.startsWith("Bearer foo"))
+        kotlin.test.assertTrue(headers["X-Api-Key"]!![0]!!.startsWith("foo"))
     }
 
     @Test
     fun interceptWithNoToken() {
-        doReturn(null).whenever(tokenProvider).getToken()
+        doReturn(null).whenever(apiKeyProvider).getApiKey()
 
         interceptor.intercept(request, ByteArray(10), exec)
 
-        assertNull(headers["Authorization"])
+        kotlin.test.assertNull(headers["X-Api-Key"])
     }
 }